+++ lib/conditional_validation.rb
@@ -1,3 +1,4 @@
+require "conditional_validation/validation_flag"
 require "conditional_validation/validation_accessor"
 
 module ConditionalValidation