+++ test/dummy/db/migrate/20131119215801_create_users.rb
@@ -2,8 +2,6 @@ class CreateUsers < ActiveRecord::Migration
   def change
     create_table :users do |t|
       t.string :name
-      t.string :address
-      t.string :city
 
       t.timestamps
     end