+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ### Unreleased
 
+* Deprecate `validation_accessor` macro in favor of `validation_flag`.
 * Documentation updates